Backend/requirements.txt
# Flask and web framework dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7

# AWS SDK and cloud services
boto3==1.34.0
botocore==1.34.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data validation and serialization
pydantic==2.5.2
marshmallow==3.20.2

# HTTP and API utilities
requests==2.31.0
urllib3>=1.25.4,<2.1

# Rate limiting and caching
Flask-Limiter==3.5.0
redis==5.0.1

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Testing dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0

# Development dependencies
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Security
cryptography==41.0.7
bcrypt==4.1.2